requirements.txt
fastapi
uvicorn
pydantic>=2.0
openai
python-dotenv
httpx
pyyaml
